+++ doc/conf.py
@@ -50,6 +50,7 @@ rst_prolog = f"""
 pygments_style = "sphinx"
 todo_include_todos = True
 toctree_collapse = False
+suppress_warnings = ["autosectionlabel.changelog"]
 
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),