+++ scim2_cli/create.py
@@ -1,11 +1,11 @@
 import click
 from click import ClickException
 from pydanclick import from_pydantic
-from scim2_client import SCIMClientError
 from scim2_models import Context
 from sphinx_click.rst_to_ansi_formatter import make_rst_to_ansi_formatter
 
 from scim2_cli.utils import DOC_URL
+from scim2_cli.utils import SCIM_EXCEPTIONS
 from scim2_cli.utils import ModelCommand
 from scim2_cli.utils import exception_to_click_error
 from scim2_cli.utils import formatted_payload
@@ -16,7 +16,7 @@ def create_payload(client, payload, indent):
     try:
         response = client.create(payload, raise_scim_errors=False)
 
-    except SCIMClientError as scim_exc:
+    except SCIM_EXCEPTIONS as scim_exc:
         raise exception_to_click_error(scim_exc) from scim_exc
 
     payload = formatted_payload(response.model_dump(), indent)