+++ scim2_cli/replace.py
@@ -1,13 +1,13 @@
 import click
 from click import ClickException
 from pydanclick import from_pydantic
-from scim2_client import SCIMClientError
 from scim2_models import Context
 from sphinx_click.rst_to_ansi_formatter import make_rst_to_ansi_formatter
 
 from scim2_cli.utils import exception_to_click_error
 
 from .utils import DOC_URL
+from .utils import SCIM_EXCEPTIONS
 from .utils import ModelCommand
 from .utils import formatted_payload
 from .utils import unacceptable_fields
@@ -17,7 +17,7 @@ def replace_payload(client, payload, indent):
     try:
         response = client.replace(payload, raise_scim_errors=False)
 
-    except SCIMClientError as scim_exc:
+    except SCIM_EXCEPTIONS as scim_exc:
         raise exception_to_click_error(scim_exc) from scim_exc
 
     payload = formatted_payload(response.model_dump(), indent)