+++ scim2_cli/search.py
@@ -1,11 +1,11 @@
 import click
-from scim2_client import SCIMClientError
 from scim2_models import SearchRequest
 from sphinx_click.rst_to_ansi_formatter import make_rst_to_ansi_formatter
 
 from scim2_cli.utils import exception_to_click_error
 
 from .utils import DOC_URL
+from .utils import SCIM_EXCEPTIONS
 from .utils import formatted_payload
 
 
@@ -91,7 +91,7 @@ def search_cli(
             raise_scim_errors=False,
         )
 
-    except SCIMClientError as scim_exc:
+    except SCIM_EXCEPTIONS as scim_exc:
         raise exception_to_click_error(scim_exc) from scim_exc
 
     payload = formatted_payload(response.model_dump(), indent)